requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
langchain==0.1.4
langchain-community==0.0.19
gTTS==2.5.0
SpeechRecognition==3.10.1
python-multipart==0.0.7
python-dotenv==1.0.1
